# docs/swagger.yaml
basePath: /
definitions:
...
  domain.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
  payload.CreateRoomType:
    properties:
      accommodate:
        type: integer
      detail:
        type: string
      isActive:
        type: boolean
      name:
        type: string
      price:
        type: number
      room:
        items:
          $ref: '#/definitions/payload.RoomOnCreateRoomType'
        type: array
    required:
    - accommodate
    - detail
    - name
    - price
    type: object
...
  payload.RoomOnCreateRoomType:
    properties:
      isActive:
        type: boolean
      roomNumber:
        type: string
    required:
    - roomNumber
    type: object
...
  payload.ServiceCreateDTO:
    properties:
      isActive:
        type: boolean
      name:
        type: string
      price:
        type: number
      serviceTypeId:
        type: integer
    required:
    - name
    - price
    - serviceTypeId
    type: object
  payload.ServiceTypeCreateDTO:
    properties:
      detail:
        type: string
      isActive:
        type: boolean
      name:
        type: string
    required:
    - detail
    - isActive
    - name
    type: object
  payload.ServiceTypeUpdateDTO:
    properties:
      detail:
        type: string
      id:
        type: integer
      isActive:
        type: boolean
      name:
        type: string
    required:
    - detail
    - id
    - name
    type: object
  payload.ServiceTypeUpdateIsActiveDTO:
    properties:
      id:
        type: integer
      isActive:
        type: boolean
    required:
    - id
    type: object
  payload.ServiceUpdateDTO:
    properties:
      id:
        type: integer
      isActive:
        type: boolean
      name:
        type: string
      price:
        type: number
      serviceTypeId:
        type: integer
    required:
    - id
    - name
    - price
    - serviceTypeId
    type: object
  payload.ServiceUpdateIsActiveDTO:
    properties:
      id:
        type: integer
      isActive:
        type: boolean
    required:
    - id
    type: object
  payload.UpdateRoomType:
    properties:
      accommodate:
        type: integer
      detail:
        type: string
      id:
        type: integer
      isActive:
        type: boolean
      name:
        type: string
      price:
        type: number
    required:
    - accommodate
    - detail
    - id
    - name
    - price
    type: object
  payload.UpdateRoomTypeIsActiveDTO:
    properties:
      id:
        type: integer
      isActive:
        type: boolean
    required:
    - id
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is a sample server celler server.
  title: CPE241 Project Backend API
  version: "1.0"
paths:
...
  /api/room_type/:
    post:
      consumes:
      - application/json
      description: Create room type
      parameters:
      - description: Room Type
        in: body
        name: roomType
        required: true
        schema:
          $ref: '#/definitions/payload.CreateRoomType'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
      summary: Create room type
      tags:
      - room_type
    put:
      consumes:
      - application/json
      description: Update room type
      parameters:
      - description: Payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/payload.UpdateRoomType'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
      summary: Update room type
      tags:
      - room_type
  /api/room_type/{id}:
    get:
      consumes:
      - application/json
      description: Get room type by id
      parameters:
      - description: Room Type ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
      summary: Get room type by id
      tags:
      - room_type
  /api/room_type/active:
    post:
      consumes:
      - application/json
      description: Update room type is active
      parameters:
      - description: Payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/payload.UpdateRoomTypeIsActiveDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
      summary: Update room type is active
      tags:
      - room_type
...
  /api/service/:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
      summary: Get all services
      tags:
      - service
    post:
      consumes:
      - application/json
      parameters:
      - description: Service
        in: body
        name: service
        required: true
        schema:
          $ref: '#/definitions/payload.ServiceCreateDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
      summary: Create service
      tags:
      - service
    put:
      consumes:
      - application/json
      parameters:
      - description: Service
        in: body
        name: service
        required: true
        schema:
          $ref: '#/definitions/payload.ServiceUpdateDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
      summary: Update service
      tags:
      - service
  /api/service/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Service ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
      summary: Get service by ID
      tags:
      - service
  /api/service/active:
    put:
      consumes:
      - application/json
      parameters:
      - description: Service
        in: body
        name: service
        required: true
        schema:
          $ref: '#/definitions/payload.ServiceUpdateIsActiveDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
      summary: Update service is active
      tags:
      - service
  /api/service_type:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
      summary: Get all service types
      tags:
      - service_type
    post:
      consumes:
      - application/json
      parameters:
      - description: Service Type
        in: body
        name: serviceType
        required: true
        schema:
          $ref: '#/definitions/payload.ServiceTypeCreateDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
      summary: Create service type
      tags:
      - service_type
  /api/service_type/:
    put:
      consumes:
      - application/json
      parameters:
      - description: Service Type ID
        in: path
        name: id
        required: true
        type: integer
      - description: Service Type
        in: body
        name: serviceType
        required: true
        schema:
          $ref: '#/definitions/payload.ServiceTypeUpdateDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
      summary: Update service type
      tags:
      - service_type
  /api/service_type/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Service Type ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
      summary: Get service type by ID
      tags:
      - service_type
  /api/service_type/active:
    post:
      consumes:
      - application/json
      parameters:
      - description: Service Type
        in: body
        name: isActive
        required: true
        schema:
          $ref: '#/definitions/payload.ServiceTypeUpdateIsActiveDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
      summary: Update service type is active
      tags:
      - service_type
...
schemes:
- https
- http
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"